[CCUBE-1298][SR] Renamed the render-toast file to toasts and renamed RenderToasts to Toasts, undid the changes done to the toast file.
