+++ src/components/common/toast/index.ts
@@ -1,2 +1,2 @@
+export * from "./toast";
 export * from "./toasts";
-export * from "./render-toasts";
