+++ src/form-builder.tsx
@@ -1,5 +1,5 @@
 import { MainPanel, SidePanel } from "./components";
-import { RenderToasts } from "./components/common";
+import { Toasts } from "./components/common";
 import { DisplayProvider } from "./context-providers";
 import { BuilderProvider } from "./context-providers/builder";
 import { Container, ToastWrapper, Wrapper } from "./form-builder.styles";
@@ -11,7 +11,7 @@ export const FormBuilder = () => {
                 <Wrapper>
                     <Container type="grid" stretch>
                         <ToastWrapper>
-                            <RenderToasts />
+                            <Toasts />
                         </ToastWrapper>
                         <MainPanel />
                         <SidePanel />
